Significant laser frequency seek modification
This code seems to work the best of all methods I have tried. It relies
on a simple while loop that chooses a stepsize based on the distance
away and a direction based on whether the frequency is higher or lower
than desired. I can reliably obtain about ~3 GHz accuracies within about
30-60 seconds, depending on how far away I am.