Single mode finding routine update
Routine now "wraps around" from the current value of the laser current,
instead of always starting at the 0'th element. This is based on the
principle that a good current value is likely to be near the one that we
already found. The sign of the movement might need to be changed,
however.