Toptica utilities
These are some basic functions I wrote with the aim of improving the
mode hop free range of the laser in an automated fashion. It relies on
using a Fabry Perot cavity with a low FSR in passive mode, i.e. where
the cavity length is fixed, and using the analog mode of the DL Pro to
tee off the piezo scan signal.